+++ resources/views/bootstrap-4/includes/bulk-select-row.blade.php
@@ -1,4 +1,4 @@
-@if ($bulkActionsEnabled && count($bulkActions) && (($selectPage && $rows->total() > $rows->count()) || count($selected)))
+@if ($bulkActionsEnabled && count($this->bulkActions) && (($selectPage && $rows->total() > $rows->count()) || count($selected)))
     <x-livewire-tables::bs4.table.row wire:key="row-message">
         <x-livewire-tables::bs4.table.cell colspan="{{ $colspan }}">
             @if (count($selected) && !$selectAll && !$selectPage)