+++ resources/views/bootstrap-5/components/table/footer.blade.php
@@ -1,0 +1,5 @@
+@props(['text' => null, 'customAttributes' => []])
+
+<td {{ $attributes->merge($customAttributes) }}>
+    {{ $text ?? $slot }}
+</td>