+++ resources/views/bootstrap-5/components/table/row-columns.blade.php
@@ -7,7 +7,7 @@
             :id="method_exists($this, 'setTableDataId') ? $this->setTableDataId($column, $row) : ''"
             :customAttributes="method_exists($this, 'setTableDataAttributes') ? $this->setTableDataAttributes($column, $row) : []"
         >
-            @if ($column->asHtml)
+            @if ($column->isHtml())
                 {{ new \Illuminate\Support\HtmlString($column->formatted($row)) }}
             @else
                 {{ $column->formatted($row) }}