+++ resources/views/bootstrap-5/components/table/table.blade.php
@@ -1,4 +1,4 @@
-<div class="table-responsive">
+<div class="{{ $this->responsive ? 'table-responsive' : '' }}">
     <table {{ $attributes->except('wire:sortable') }} class="table table-striped">
         <thead>
             <tr>
@@ -9,5 +9,17 @@
         <tbody {{ $attributes->only('wire:sortable') }}>
             {{ $body }}
         </tbody>
+
+        @if ($useHeaderAsFooter || $customFooter)
+            <tfoot>
+                @if ($useHeaderAsFooter)
+                    <tr>
+                        {{ $head }}
+                    </tr>
+                @elseif($customFooter)
+                    {{ $foot }}
+                @endif
+            </tfoot>
+        @endif
     </table>
 </div>