+++ src/Views/Column.php
@@ -2,6 +2,7 @@
 
 namespace Rappasoft\LaravelLivewireTables\Views;
 
+use Illuminate\Support\HtmlString;
 use Illuminate\Support\Str;
 
 /**
@@ -78,6 +79,26 @@ class Column
      * @var bool
      */
     public bool $selected = false;
+
+    /**
+     * @var bool
+     */
+    public bool $footer = false;
+
+    /**
+     * @var
+     */
+    public $footerCallback;
+
+    /**
+     * @var
+     */
+    public $linkCallback;
+
+    /**
+     * @var ?string
+     */
+    public ?string $linkTarget;
 
     /**
      * Column constructor.
@@ -203,6 +224,14 @@ class Column
     }
 
     /**
+     * @return bool
+     */
+    public function isHtml(): bool
+    {
+        return $this->asHtml === true;
+    }
+
+    /**
      * @return string|null
      */
     public function class(): ?string
@@ -266,6 +295,15 @@ class Column
 
         if ($this->formatCallback) {
             $value = call_user_func($this->formatCallback, $value, $column, $row);
+        }
+
+        if ($this->linkCallback) {
+            $url = call_user_func($this->linkCallback, $value, $column, $row);
+
+            if ($url) {
+                $linkTarget = $this->linkTarget ? "target='$this->linkTarget'" : '';
+                $value = new HtmlString("<a href='$url' $linkTarget>$value</a>");
+            }
         }
 
         return $value;
@@ -357,5 +395,55 @@ class Column
     public function isSelected(): bool
     {
         return $this->selected;
+    }
+
+    /**
+     * @return bool
+     */
+    public function hasFooter(): bool
+    {
+        return $this->footer === true;
+    }
+
+    /**
+     * @return $this
+     */
+    public function footer($callback = null): self
+    {
+        $this->footer = true;
+
+        $this->footerCallback = $callback;
+
+        return $this;
+    }
+
+    /**
+     * @param $rows
+     *
+     * @return false|mixed|null
+     */
+    public function footerFormatted($rows)
+    {
+        $value = null;
+
+        if ($this->footerCallback) {
+            $value = call_user_func($this->footerCallback, $rows);
+        }
+
+        return $value;
+    }
+
+    /**
+     * @param  callable  $callable
+     * @param  string|null  $target
+     *
+     * @return $this
+     */
+    public function linkTo(callable $callable, string $target = null): self
+    {
+        $this->linkCallback = $callable;
+        $this->linkTarget = $target;
+
+        return $this;
     }
 }